+++ CHANGELOG.md
@@ -50,6 +50,9 @@
   snapshots instead of moving or copying them. This can be invoked with
   `--import-snapshot <path> --import-mode=symlink`.
 
+- [#4533](https://github.com/ChainSafe/forest/pull/4641) Added `build_info`
+  metric to Prometheus metrics, which include the current build's version.
+
 - [#4628](https://github.com/ChainSafe/forest/issues/4628) Added support for
   devnets (2k networks) in the offline Forest.